helpers made, classes made, framework check added
